boot: enable the A20 line
This commit adds Fast A20 Gate (enabling A20 line).

[Short Introduction]
The A20 Address Line is the physical representation of the 21st bit
(number 20, counting from 0) of any memory access. When the IBM-AT
(Intel 286) was introduced, it was able to access up to sixteen
megabytes of memory (instead of the 1 MByte of the 8086). But to remain
compatible with the 8086, a quirk in the 8086 architecture (memory
wraparound) had to be duplicated in the AT. To achieve this, the A20
line on the address bus was disabled by default.

Link: https://wiki.osdev.org/A20_Line
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>